LayoutTests/imported/w3c/web-platform-tests/webrtc/interfaces.https-expected.txt

CONSOLE MESSAGE: line 309: callback not yet supported
CONSOLE MESSAGE: line 309: callback not yet supported
CONSOLE MESSAGE: line 309: callback not yet supported
CONSOLE MESSAGE: line 309: callback not yet supported

PASS Main test driver 
FAIL Test driver for asyncInitCertificate undefined is not a function (near '...RTCPeerConnection.generateCertificate...')
FAIL Test driver for asyncInitTransports assert_unreached: Failed to run asyncInitTransports: ReferenceError: Can't find variable: RTCSctpTransport Reached unreachable code
PASS Test driver for asyncInitMediaStreamTrack 
FAIL MediaStreamTrack interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS MediaStreamTrack interface object length 
PASS MediaStreamTrack interface object name 
FAIL MediaStreamTrack interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS MediaStreamTrack interface: existence and properties of interface prototype object's "constructor" property 
PASS MediaStreamTrack interface: existence and properties of interface prototype object's @@unscopables property 
FAIL MediaStreamTrack interface: attribute isolated assert_true: The prototype object must have a property "isolated" expected true got false
PASS Unscopable handled correctly for isolated property on MediaStreamTrack 
FAIL MediaStreamTrack interface: attribute onisolationchange assert_true: The prototype object must have a property "onisolationchange" expected true got false
PASS Unscopable handled correctly for onisolationchange property on MediaStreamTrack 
PASS MediaStreamTrack must be primary interface of idlTestObjects.mediaStreamTrack 
FAIL Stringification of idlTestObjects.mediaStreamTrack undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "isolated" with the proper type assert_inherits: property "isolated" not found in prototype chain
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "onisolationchange" with the proper type assert_inherits: property "onisolationchange" not found in prototype chain
PASS MediaStreamTrack must be primary interface of generateMediaStreamTrack('audio') 
FAIL Stringification of generateMediaStreamTrack('audio') undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "isolated" with the proper type assert_inherits: property "isolated" not found in prototype chain
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "onisolationchange" with the proper type assert_inherits: property "onisolationchange" not found in prototype chain
FAIL RTCPeerConnection interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCPeerConnection interface object length 
PASS RTCPeerConnection interface object name 
FAIL RTCPeerConnection interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCPeerConnection interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCPeerConnection interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCPeerConnection interface: operation createOffer(RTCOfferOptions) 
PASS Unscopable handled correctly for createOffer(RTCOfferOptions) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation createAnswer(RTCAnswerOptions) 
PASS Unscopable handled correctly for createAnswer(RTCAnswerOptions) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit) assert_equals: property has wrong .length expected 1 but got 0
PASS Unscopable handled correctly for setLocalDescription(RTCSessionDescriptionInit) on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute localDescription 
PASS Unscopable handled correctly for localDescription property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute currentLocalDescription 
PASS Unscopable handled correctly for currentLocalDescription property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute pendingLocalDescription 
PASS Unscopable handled correctly for pendingLocalDescription property on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit) assert_equals: property has wrong .length expected 1 but got 0
PASS Unscopable handled correctly for setRemoteDescription(RTCSessionDescriptionInit) on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute remoteDescription 
PASS Unscopable handled correctly for remoteDescription property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute currentRemoteDescription 
PASS Unscopable handled correctly for currentRemoteDescription property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute pendingRemoteDescription 
PASS Unscopable handled correctly for pendingRemoteDescription property on RTCPeerConnection 
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object]) 
PASS Unscopable handled correctly for addIceCandidate([object Object],[object Object]) on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute signalingState 
PASS Unscopable handled correctly for signalingState property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute iceGatheringState 
PASS Unscopable handled correctly for iceGatheringState property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute iceConnectionState 
PASS Unscopable handled correctly for iceConnectionState property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute connectionState 
PASS Unscopable handled correctly for connectionState property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute canTrickleIceCandidates assert_true: The prototype object must have a property "canTrickleIceCandidates" expected true got false
PASS Unscopable handled correctly for canTrickleIceCandidates property on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation getDefaultIceServers() assert_own_property: interface object missing static operation expected property "getDefaultIceServers" missing
PASS Unscopable handled correctly for getDefaultIceServers() on RTCPeerConnection 
PASS RTCPeerConnection interface: operation getConfiguration() 
PASS Unscopable handled correctly for getConfiguration() on RTCPeerConnection 
PASS RTCPeerConnection interface: operation setConfiguration(RTCConfiguration) 
PASS Unscopable handled correctly for setConfiguration(RTCConfiguration) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation close() 
PASS Unscopable handled correctly for close() on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute onnegotiationneeded 
PASS Unscopable handled correctly for onnegotiationneeded property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute onicecandidate 
PASS Unscopable handled correctly for onicecandidate property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute onicecandidateerror assert_true: The prototype object must have a property "onicecandidateerror" expected true got false
PASS Unscopable handled correctly for onicecandidateerror property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute onsignalingstatechange 
PASS Unscopable handled correctly for onsignalingstatechange property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute oniceconnectionstatechange 
PASS Unscopable handled correctly for oniceconnectionstatechange property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute onicegatheringstatechange 
PASS Unscopable handled correctly for onicegatheringstatechange property on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute onconnectionstatechange 
PASS Unscopable handled correctly for onconnectionstatechange property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute onfingerprintfailure assert_true: The prototype object must have a property "onfingerprintfailure" expected true got false
PASS Unscopable handled correctly for onfingerprintfailure property on RTCPeerConnection 
PASS RTCPeerConnection interface: operation createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) 
PASS Unscopable handled correctly for createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 1 but got 0
PASS Unscopable handled correctly for setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) 
PASS Unscopable handled correctly for createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 1 but got 0
PASS Unscopable handled correctly for setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) 
PASS Unscopable handled correctly for addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation generateCertificate(AlgorithmIdentifier) assert_own_property: interface object missing static operation expected property "generateCertificate" missing
PASS Unscopable handled correctly for generateCertificate(AlgorithmIdentifier) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation getSenders() 
PASS Unscopable handled correctly for getSenders() on RTCPeerConnection 
PASS RTCPeerConnection interface: operation getReceivers() 
PASS Unscopable handled correctly for getReceivers() on RTCPeerConnection 
PASS RTCPeerConnection interface: operation getTransceivers() 
PASS Unscopable handled correctly for getTransceivers() on RTCPeerConnection 
PASS RTCPeerConnection interface: operation addTrack(MediaStreamTrack, MediaStream) 
PASS Unscopable handled correctly for addTrack(MediaStreamTrack, MediaStream) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation removeTrack(RTCRtpSender) 
PASS Unscopable handled correctly for removeTrack(RTCRtpSender) on RTCPeerConnection 
PASS RTCPeerConnection interface: operation addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) 
PASS Unscopable handled correctly for addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute ontrack 
PASS Unscopable handled correctly for ontrack property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute sctp assert_true: The prototype object must have a property "sctp" expected true got false
PASS Unscopable handled correctly for sctp property on RTCPeerConnection 
PASS RTCPeerConnection interface: operation createDataChannel(USVString, RTCDataChannelInit) 
PASS Unscopable handled correctly for createDataChannel(USVString, RTCDataChannelInit) on RTCPeerConnection 
PASS RTCPeerConnection interface: attribute ondatachannel 
PASS Unscopable handled correctly for ondatachannel property on RTCPeerConnection 
PASS RTCPeerConnection interface: operation getStats(MediaStreamTrack) 
PASS Unscopable handled correctly for getStats(MediaStreamTrack) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation setIdentityProvider(DOMString, RTCIdentityProviderOptions) assert_own_property: interface prototype object missing non-static operation expected property "setIdentityProvider" missing
PASS Unscopable handled correctly for setIdentityProvider(DOMString, RTCIdentityProviderOptions) on RTCPeerConnection 
FAIL RTCPeerConnection interface: operation getIdentityAssertion() assert_own_property: interface prototype object missing non-static operation expected property "getIdentityAssertion" missing
PASS Unscopable handled correctly for getIdentityAssertion() on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute peerIdentity assert_true: The prototype object must have a property "peerIdentity" expected true got false
PASS Unscopable handled correctly for peerIdentity property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute idpLoginUrl assert_true: The prototype object must have a property "idpLoginUrl" expected true got false
PASS Unscopable handled correctly for idpLoginUrl property on RTCPeerConnection 
FAIL RTCPeerConnection interface: attribute idpErrorInfo assert_true: The prototype object must have a property "idpErrorInfo" expected true got false
PASS Unscopable handled correctly for idpErrorInfo property on RTCPeerConnection 
PASS RTCPeerConnection must be primary interface of new RTCPeerConnection() 
FAIL Stringification of new RTCPeerConnection() undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCOfferOptions)" with the proper type 
PASS RTCPeerConnection interface: calling createOffer(RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCAnswerOptions)" with the proper type 
PASS RTCPeerConnection interface: calling createAnswer(RTCAnswerOptions) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit)" with the proper type 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "localDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentLocalDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingLocalDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit)" with the proper type 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "remoteDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentRemoteDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingRemoteDescription" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object])" with the proper type 
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object]) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "signalingState" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceGatheringState" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceConnectionState" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "connectionState" with the proper type 
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "canTrickleIceCandidates" with the proper type assert_inherits: property "canTrickleIceCandidates" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getDefaultIceServers()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getConfiguration()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setConfiguration(RTCConfiguration)" with the proper type 
PASS RTCPeerConnection interface: calling setConfiguration(RTCConfiguration) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "close()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onnegotiationneeded" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidate" with the proper type 
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidateerror" with the proper type assert_inherits: property "onicecandidateerror" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onsignalingstatechange" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "oniceconnectionstatechange" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicegatheringstatechange" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onconnectionstatechange" with the proper type 
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onfingerprintfailure" with the proper type assert_inherits: property "onfingerprintfailure" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions)" with the proper type 
PASS RTCPeerConnection interface: calling createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback)" with the proper type 
PASS RTCPeerConnection interface: calling createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type 
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "generateCertificate(AlgorithmIdentifier)" with the proper type 
FAIL RTCPeerConnection interface: calling generateCertificate(AlgorithmIdentifier) on new RTCPeerConnection() with too few arguments must throw TypeError assert_own_property: interface object must have static operation as own property expected property "generateCertificate" missing
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getSenders()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getReceivers()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getTransceivers()" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTrack(MediaStreamTrack, MediaStream)" with the proper type 
PASS RTCPeerConnection interface: calling addTrack(MediaStreamTrack, MediaStream) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "removeTrack(RTCRtpSender)" with the proper type 
PASS RTCPeerConnection interface: calling removeTrack(RTCRtpSender) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTransceiver([object Object],[object Object], RTCRtpTransceiverInit)" with the proper type 
PASS RTCPeerConnection interface: calling addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ontrack" with the proper type 
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "sctp" with the proper type assert_inherits: property "sctp" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createDataChannel(USVString, RTCDataChannelInit)" with the proper type 
PASS RTCPeerConnection interface: calling createDataChannel(USVString, RTCDataChannelInit) on new RTCPeerConnection() with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ondatachannel" with the proper type 
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getStats(MediaStreamTrack)" with the proper type 
PASS RTCPeerConnection interface: calling getStats(MediaStreamTrack) on new RTCPeerConnection() with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setIdentityProvider(DOMString, RTCIdentityProviderOptions)" with the proper type assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: calling setIdentityProvider(DOMString, RTCIdentityProviderOptions) on new RTCPeerConnection() with too few arguments must throw TypeError assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getIdentityAssertion()" with the proper type assert_inherits: property "getIdentityAssertion" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "peerIdentity" with the proper type assert_inherits: property "peerIdentity" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpLoginUrl" with the proper type assert_inherits: property "idpLoginUrl" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpErrorInfo" with the proper type assert_inherits: property "idpErrorInfo" not found in prototype chain
PASS RTCSessionDescription interface: existence and properties of interface object 
PASS RTCSessionDescription interface object length 
PASS RTCSessionDescription interface object name 
PASS RTCSessionDescription interface: existence and properties of interface prototype object 
PASS RTCSessionDescription interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCSessionDescription interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCSessionDescription interface: attribute type 
PASS Unscopable handled correctly for type property on RTCSessionDescription 
PASS RTCSessionDescription interface: attribute sdp 
PASS Unscopable handled correctly for sdp property on RTCSessionDescription 
PASS RTCSessionDescription interface: operation toJSON() 
PASS Unscopable handled correctly for toJSON() on RTCSessionDescription 
PASS RTCSessionDescription must be primary interface of new RTCSessionDescription({ type: 'offer' }) 
PASS Stringification of new RTCSessionDescription({ type: 'offer' }) 
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "type" with the proper type 
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "sdp" with the proper type 
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "toJSON()" with the proper type 
PASS Test default toJSON operation of RTCSessionDescription 
PASS RTCIceCandidate interface: existence and properties of interface object 
FAIL RTCIceCandidate interface object length assert_equals: wrong value for RTCIceCandidate.length expected 0 but got 1
PASS RTCIceCandidate interface object name 
PASS RTCIceCandidate interface: existence and properties of interface prototype object 
PASS RTCIceCandidate interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCIceCandidate interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCIceCandidate interface: attribute candidate 
PASS Unscopable handled correctly for candidate property on RTCIceCandidate 
PASS RTCIceCandidate interface: attribute sdpMid 
PASS Unscopable handled correctly for sdpMid property on RTCIceCandidate 
PASS RTCIceCandidate interface: attribute sdpMLineIndex 
PASS Unscopable handled correctly for sdpMLineIndex property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute foundation assert_true: The prototype object must have a property "foundation" expected true got false
PASS Unscopable handled correctly for foundation property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute component assert_true: The prototype object must have a property "component" expected true got false
PASS Unscopable handled correctly for component property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute priority assert_true: The prototype object must have a property "priority" expected true got false
PASS Unscopable handled correctly for priority property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute ip assert_true: The prototype object must have a property "ip" expected true got false
PASS Unscopable handled correctly for ip property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute protocol assert_true: The prototype object must have a property "protocol" expected true got false
PASS Unscopable handled correctly for protocol property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute port assert_true: The prototype object must have a property "port" expected true got false
PASS Unscopable handled correctly for port property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute type assert_true: The prototype object must have a property "type" expected true got false
PASS Unscopable handled correctly for type property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute tcpType assert_true: The prototype object must have a property "tcpType" expected true got false
PASS Unscopable handled correctly for tcpType property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute relatedAddress assert_true: The prototype object must have a property "relatedAddress" expected true got false
PASS Unscopable handled correctly for relatedAddress property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute relatedPort assert_true: The prototype object must have a property "relatedPort" expected true got false
PASS Unscopable handled correctly for relatedPort property on RTCIceCandidate 
FAIL RTCIceCandidate interface: attribute usernameFragment assert_true: The prototype object must have a property "usernameFragment" expected true got false
PASS Unscopable handled correctly for usernameFragment property on RTCIceCandidate 
PASS RTCIceCandidate interface: operation toJSON() 
PASS Unscopable handled correctly for toJSON() on RTCIceCandidate 
PASS RTCIceCandidate must be primary interface of new RTCIceCandidate({ sdpMid: 1 }) 
PASS Stringification of new RTCIceCandidate({ sdpMid: 1 }) 
PASS RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "candidate" with the proper type 
PASS RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMid" with the proper type 
PASS RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMLineIndex" with the proper type 
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "foundation" with the proper type assert_inherits: property "foundation" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "component" with the proper type assert_inherits: property "component" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "priority" with the proper type assert_inherits: property "priority" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "ip" with the proper type assert_inherits: property "ip" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "protocol" with the proper type assert_inherits: property "protocol" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "port" with the proper type assert_inherits: property "port" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "type" with the proper type assert_inherits: property "type" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "tcpType" with the proper type assert_inherits: property "tcpType" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedAddress" with the proper type assert_inherits: property "relatedAddress" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedPort" with the proper type assert_inherits: property "relatedPort" not found in prototype chain
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "usernameFragment" with the proper type assert_inherits: property "usernameFragment" not found in prototype chain
PASS RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "toJSON()" with the proper type 
PASS Test toJSON operation of RTCIceCandidate 
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface object length assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface object name assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
FAIL RTCPeerConnectionIceEvent interface: attribute candidate assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
PASS Unscopable handled correctly for candidate property on RTCPeerConnectionIceEvent 
FAIL RTCPeerConnectionIceEvent interface: attribute url assert_own_property: self does not have own property "RTCPeerConnectionIceEvent" expected property "RTCPeerConnectionIceEvent" missing
PASS Unscopable handled correctly for url property on RTCPeerConnectionIceEvent 
FAIL RTCPeerConnectionIceEvent must be primary interface of new RTCPeerConnectionIceEvent('ice') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL Stringification of new RTCPeerConnectionIceEvent('ice') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "candidate" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "url" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object length assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object name assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute hostCandidate assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
PASS Unscopable handled correctly for hostCandidate property on RTCPeerConnectionIceErrorEvent 
FAIL RTCPeerConnectionIceErrorEvent interface: attribute url assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
PASS Unscopable handled correctly for url property on RTCPeerConnectionIceErrorEvent 
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorCode assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
PASS Unscopable handled correctly for errorCode property on RTCPeerConnectionIceErrorEvent 
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorText assert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
PASS Unscopable handled correctly for errorText property on RTCPeerConnectionIceErrorEvent 
FAIL RTCPeerConnectionIceErrorEvent must be primary interface of new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL Stringification of new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "hostCandidate" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "url" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "errorCode" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "errorText" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCPeerConnectionIceErrorEvent"
FAIL RTCCertificate interface: existence and properties of interface object assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface object length assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface object name assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
FAIL RTCCertificate interface: attribute expires assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
PASS Unscopable handled correctly for expires property on RTCCertificate 
FAIL RTCCertificate interface: operation getFingerprints() assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
PASS Unscopable handled correctly for getFingerprints() on RTCCertificate 
FAIL RTCCertificate interface: operation getAlgorithm() assert_own_property: self does not have own property "RTCCertificate" expected property "RTCCertificate" missing
PASS Unscopable handled correctly for getAlgorithm() on RTCCertificate 
FAIL RTCCertificate must be primary interface of idlTestObjects.certificate assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.certificate assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCCertificate interface: idlTestObjects.certificate must inherit property "expires" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCCertificate interface: idlTestObjects.certificate must inherit property "getFingerprints()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCCertificate interface: idlTestObjects.certificate must inherit property "getAlgorithm()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
PASS RTCRtpSender interface: existence and properties of interface object 
PASS RTCRtpSender interface object length 
PASS RTCRtpSender interface object name 
PASS RTCRtpSender interface: existence and properties of interface prototype object 
PASS RTCRtpSender interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpSender interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCRtpSender interface: attribute track 
PASS Unscopable handled correctly for track property on RTCRtpSender 
FAIL RTCRtpSender interface: attribute transport assert_true: The prototype object must have a property "transport" expected true got false
PASS Unscopable handled correctly for transport property on RTCRtpSender 
FAIL RTCRtpSender interface: attribute rtcpTransport assert_true: The prototype object must have a property "rtcpTransport" expected true got false
PASS Unscopable handled correctly for rtcpTransport property on RTCRtpSender 
FAIL RTCRtpSender interface: operation getCapabilities(DOMString) assert_own_property: interface object missing static operation expected property "getCapabilities" missing
PASS Unscopable handled correctly for getCapabilities(DOMString) on RTCRtpSender 
FAIL RTCRtpSender interface: operation setParameters(RTCRtpParameters) assert_own_property: interface prototype object missing non-static operation expected property "setParameters" missing
PASS Unscopable handled correctly for setParameters(RTCRtpParameters) on RTCRtpSender 
PASS RTCRtpSender interface: operation getParameters() 
PASS Unscopable handled correctly for getParameters() on RTCRtpSender 
PASS RTCRtpSender interface: operation replaceTrack(MediaStreamTrack) 
PASS Unscopable handled correctly for replaceTrack(MediaStreamTrack) on RTCRtpSender 
FAIL RTCRtpSender interface: operation getStats() assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
PASS Unscopable handled correctly for getStats() on RTCRtpSender 
FAIL RTCRtpSender interface: attribute dtmf assert_true: The prototype object must have a property "dtmf" expected true got false
PASS Unscopable handled correctly for dtmf property on RTCRtpSender 
PASS RTCRtpSender must be primary interface of new RTCPeerConnection().addTransceiver('audio').sender 
PASS Stringification of new RTCPeerConnection().addTransceiver('audio').sender 
PASS RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "track" with the proper type 
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "transport" with the proper type assert_inherits: property "transport" not found in prototype chain
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "rtcpTransport" with the proper type assert_inherits: property "rtcpTransport" not found in prototype chain
PASS RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getCapabilities(DOMString)" with the proper type 
FAIL RTCRtpSender interface: calling getCapabilities(DOMString) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeError assert_own_property: interface object must have static operation as own property expected property "getCapabilities" missing
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "setParameters(RTCRtpParameters)" with the proper type assert_inherits: property "setParameters" not found in prototype chain
FAIL RTCRtpSender interface: calling setParameters(RTCRtpParameters) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeError assert_inherits: property "setParameters" not found in prototype chain
PASS RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getParameters()" with the proper type 
PASS RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "replaceTrack(MediaStreamTrack)" with the proper type 
PASS RTCRtpSender interface: calling replaceTrack(MediaStreamTrack) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeError 
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getStats()" with the proper type assert_inherits: property "getStats" not found in prototype chain
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "dtmf" with the proper type assert_inherits: property "dtmf" not found in prototype chain
PASS RTCRtpReceiver interface: existence and properties of interface object 
PASS RTCRtpReceiver interface object length 
PASS RTCRtpReceiver interface object name 
PASS RTCRtpReceiver interface: existence and properties of interface prototype object 
PASS RTCRtpReceiver interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpReceiver interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCRtpReceiver interface: attribute track 
PASS Unscopable handled correctly for track property on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: attribute transport assert_true: The prototype object must have a property "transport" expected true got false
PASS Unscopable handled correctly for transport property on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: attribute rtcpTransport assert_true: The prototype object must have a property "rtcpTransport" expected true got false
PASS Unscopable handled correctly for rtcpTransport property on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: operation getCapabilities(DOMString) assert_own_property: interface object missing static operation expected property "getCapabilities" missing
PASS Unscopable handled correctly for getCapabilities(DOMString) on RTCRtpReceiver 
PASS RTCRtpReceiver interface: operation getParameters() 
PASS Unscopable handled correctly for getParameters() on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: operation getContributingSources() assert_own_property: interface prototype object missing non-static operation expected property "getContributingSources" missing
PASS Unscopable handled correctly for getContributingSources() on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: operation getSynchronizationSources() assert_own_property: interface prototype object missing non-static operation expected property "getSynchronizationSources" missing
PASS Unscopable handled correctly for getSynchronizationSources() on RTCRtpReceiver 
FAIL RTCRtpReceiver interface: operation getStats() assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
PASS Unscopable handled correctly for getStats() on RTCRtpReceiver 
PASS RTCRtpReceiver must be primary interface of new RTCPeerConnection().addTransceiver('audio').receiver 
PASS Stringification of new RTCPeerConnection().addTransceiver('audio').receiver 
PASS RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "track" with the proper type 
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "transport" with the proper type assert_inherits: property "transport" not found in prototype chain
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "rtcpTransport" with the proper type assert_inherits: property "rtcpTransport" not found in prototype chain
PASS RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getCapabilities(DOMString)" with the proper type 
FAIL RTCRtpReceiver interface: calling getCapabilities(DOMString) on new RTCPeerConnection().addTransceiver('audio').receiver with too few arguments must throw TypeError assert_own_property: interface object must have static operation as own property expected property "getCapabilities" missing
PASS RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getParameters()" with the proper type 
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getContributingSources()" with the proper type assert_inherits: property "getContributingSources" not found in prototype chain
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getSynchronizationSources()" with the proper type assert_inherits: property "getSynchronizationSources" not found in prototype chain
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getStats()" with the proper type assert_inherits: property "getStats" not found in prototype chain
FAIL RTCRtpContributingSource interface: existence and properties of interface object assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface object length assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface object name assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
FAIL RTCRtpContributingSource interface: attribute timestamp assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
PASS Unscopable handled correctly for timestamp property on RTCRtpContributingSource 
FAIL RTCRtpContributingSource interface: attribute source assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
PASS Unscopable handled correctly for source property on RTCRtpContributingSource 
FAIL RTCRtpContributingSource interface: attribute audioLevel assert_own_property: self does not have own property "RTCRtpContributingSource" expected property "RTCRtpContributingSource" missing
PASS Unscopable handled correctly for audioLevel property on RTCRtpContributingSource 
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface object assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object length assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object name assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute timestamp assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
PASS Unscopable handled correctly for timestamp property on RTCRtpSynchronizationSource 
FAIL RTCRtpSynchronizationSource interface: attribute source assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
PASS Unscopable handled correctly for source property on RTCRtpSynchronizationSource 
FAIL RTCRtpSynchronizationSource interface: attribute audioLevel assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
PASS Unscopable handled correctly for audioLevel property on RTCRtpSynchronizationSource 
FAIL RTCRtpSynchronizationSource interface: attribute voiceActivityFlag assert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
PASS Unscopable handled correctly for voiceActivityFlag property on RTCRtpSynchronizationSource 
PASS RTCRtpTransceiver interface: existence and properties of interface object 
PASS RTCRtpTransceiver interface object length 
PASS RTCRtpTransceiver interface object name 
PASS RTCRtpTransceiver interface: existence and properties of interface prototype object 
PASS RTCRtpTransceiver interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCRtpTransceiver interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCRtpTransceiver interface: attribute mid 
PASS Unscopable handled correctly for mid property on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: attribute sender 
PASS Unscopable handled correctly for sender property on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: attribute receiver 
PASS Unscopable handled correctly for receiver property on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: attribute stopped 
PASS Unscopable handled correctly for stopped property on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: attribute direction 
PASS Unscopable handled correctly for direction property on RTCRtpTransceiver 
FAIL RTCRtpTransceiver interface: attribute currentDirection assert_true: The prototype object must have a property "currentDirection" expected true got false
PASS Unscopable handled correctly for currentDirection property on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: operation setDirection(RTCRtpTransceiverDirection) 
PASS Unscopable handled correctly for setDirection(RTCRtpTransceiverDirection) on RTCRtpTransceiver 
PASS RTCRtpTransceiver interface: operation stop() 
PASS Unscopable handled correctly for stop() on RTCRtpTransceiver 
FAIL RTCRtpTransceiver interface: operation setCodecPreferences([object Object]) assert_own_property: interface prototype object missing non-static operation expected property "setCodecPreferences" missing
PASS Unscopable handled correctly for setCodecPreferences([object Object]) on RTCRtpTransceiver 
PASS RTCRtpTransceiver must be primary interface of new RTCPeerConnection().addTransceiver('audio') 
PASS Stringification of new RTCPeerConnection().addTransceiver('audio') 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "mid" with the proper type 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "sender" with the proper type 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "receiver" with the proper type 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "stopped" with the proper type 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "direction" with the proper type 
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "currentDirection" with the proper type assert_inherits: property "currentDirection" not found in prototype chain
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "setDirection(RTCRtpTransceiverDirection)" with the proper type 
PASS RTCRtpTransceiver interface: calling setDirection(RTCRtpTransceiverDirection) on new RTCPeerConnection().addTransceiver('audio') with too few arguments must throw TypeError 
PASS RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "stop()" with the proper type 
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "setCodecPreferences([object Object])" with the proper type assert_inherits: property "setCodecPreferences" not found in prototype chain
FAIL RTCRtpTransceiver interface: calling setCodecPreferences([object Object]) on new RTCPeerConnection().addTransceiver('audio') with too few arguments must throw TypeError assert_inherits: property "setCodecPreferences" not found in prototype chain
FAIL RTCDtlsTransport interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object length assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object name assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute transport assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
PASS Unscopable handled correctly for transport property on RTCDtlsTransport 
FAIL RTCDtlsTransport interface: attribute state assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
PASS Unscopable handled correctly for state property on RTCDtlsTransport 
FAIL RTCDtlsTransport interface: operation getRemoteCertificates() assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
PASS Unscopable handled correctly for getRemoteCertificates() on RTCDtlsTransport 
FAIL RTCDtlsTransport interface: attribute onstatechange assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
PASS Unscopable handled correctly for onstatechange property on RTCDtlsTransport 
FAIL RTCDtlsTransport must be primary interface of idlTestObjects.dtlsTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.dtlsTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "transport" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "state" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "getRemoteCertificates()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "onstatechange" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
PASS RTCIceTransport interface: existence and properties of interface object 
PASS RTCIceTransport interface object length 
PASS RTCIceTransport interface object name 
PASS RTCIceTransport interface: existence and properties of interface prototype object 
PASS RTCIceTransport interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCIceTransport interface: existence and properties of interface prototype object's @@unscopables property 
FAIL RTCIceTransport interface: attribute role assert_true: The prototype object must have a property "role" expected true got false
PASS Unscopable handled correctly for role property on RTCIceTransport 
FAIL RTCIceTransport interface: attribute component assert_true: The prototype object must have a property "component" expected true got false
PASS Unscopable handled correctly for component property on RTCIceTransport 
PASS RTCIceTransport interface: attribute state 
PASS Unscopable handled correctly for state property on RTCIceTransport 
PASS RTCIceTransport interface: attribute gatheringState 
PASS Unscopable handled correctly for gatheringState property on RTCIceTransport 
FAIL RTCIceTransport interface: operation getLocalCandidates() assert_own_property: interface prototype object missing non-static operation expected property "getLocalCandidates" missing
PASS Unscopable handled correctly for getLocalCandidates() on RTCIceTransport 
FAIL RTCIceTransport interface: operation getRemoteCandidates() assert_own_property: interface prototype object missing non-static operation expected property "getRemoteCandidates" missing
PASS Unscopable handled correctly for getRemoteCandidates() on RTCIceTransport 
FAIL RTCIceTransport interface: operation getSelectedCandidatePair() assert_own_property: interface prototype object missing non-static operation expected property "getSelectedCandidatePair" missing
PASS Unscopable handled correctly for getSelectedCandidatePair() on RTCIceTransport 
FAIL RTCIceTransport interface: operation getLocalParameters() assert_own_property: interface prototype object missing non-static operation expected property "getLocalParameters" missing
PASS Unscopable handled correctly for getLocalParameters() on RTCIceTransport 
FAIL RTCIceTransport interface: operation getRemoteParameters() assert_own_property: interface prototype object missing non-static operation expected property "getRemoteParameters" missing
PASS Unscopable handled correctly for getRemoteParameters() on RTCIceTransport 
FAIL RTCIceTransport interface: attribute onstatechange assert_true: The prototype object must have a property "onstatechange" expected true got false
PASS Unscopable handled correctly for onstatechange property on RTCIceTransport 
FAIL RTCIceTransport interface: attribute ongatheringstatechange assert_true: The prototype object must have a property "ongatheringstatechange" expected true got false
PASS Unscopable handled correctly for ongatheringstatechange property on RTCIceTransport 
FAIL RTCIceTransport interface: attribute onselectedcandidatepairchange assert_true: The prototype object must have a property "onselectedcandidatepairchange" expected true got false
PASS Unscopable handled correctly for onselectedcandidatepairchange property on RTCIceTransport 
FAIL RTCIceTransport must be primary interface of idlTestObjects.iceTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.iceTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "role" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "component" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "state" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "gatheringState" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getLocalCandidates()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getRemoteCandidates()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getSelectedCandidatePair()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getLocalParameters()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getRemoteParameters()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "onstatechange" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "ongatheringstatechange" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "onselectedcandidatepairchange" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCTrackEvent interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCTrackEvent interface object length 
PASS RTCTrackEvent interface object name 
FAIL RTCTrackEvent interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCTrackEvent interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCTrackEvent interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCTrackEvent interface: attribute receiver 
PASS Unscopable handled correctly for receiver property on RTCTrackEvent 
PASS RTCTrackEvent interface: attribute track 
PASS Unscopable handled correctly for track property on RTCTrackEvent 
PASS RTCTrackEvent interface: attribute streams 
PASS Unscopable handled correctly for streams property on RTCTrackEvent 
PASS RTCTrackEvent interface: attribute transceiver 
PASS Unscopable handled correctly for transceiver property on RTCTrackEvent 
PASS RTCTrackEvent must be primary interface of initTrackEvent() 
FAIL Stringification of initTrackEvent() undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
PASS RTCTrackEvent interface: initTrackEvent() must inherit property "receiver" with the proper type 
PASS RTCTrackEvent interface: initTrackEvent() must inherit property "track" with the proper type 
PASS RTCTrackEvent interface: initTrackEvent() must inherit property "streams" with the proper type 
PASS RTCTrackEvent interface: initTrackEvent() must inherit property "transceiver" with the proper type 
FAIL RTCSctpTransport interface: existence and properties of interface object assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object length assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object name assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: attribute transport assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
PASS Unscopable handled correctly for transport property on RTCSctpTransport 
FAIL RTCSctpTransport interface: attribute maxMessageSize assert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
PASS Unscopable handled correctly for maxMessageSize property on RTCSctpTransport 
FAIL RTCSctpTransport must be primary interface of idlTestObjects.sctpTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.sctpTransport assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCSctpTransport interface: idlTestObjects.sctpTransport must inherit property "transport" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCSctpTransport interface: idlTestObjects.sctpTransport must inherit property "maxMessageSize" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDataChannel interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCDataChannel interface object length 
PASS RTCDataChannel interface object name 
FAIL RTCDataChannel interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCDataChannel interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCDataChannel interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCDataChannel interface: attribute label 
PASS Unscopable handled correctly for label property on RTCDataChannel 
PASS RTCDataChannel interface: attribute ordered 
PASS Unscopable handled correctly for ordered property on RTCDataChannel 
PASS RTCDataChannel interface: attribute maxPacketLifeTime 
PASS Unscopable handled correctly for maxPacketLifeTime property on RTCDataChannel 
PASS RTCDataChannel interface: attribute maxRetransmits 
PASS Unscopable handled correctly for maxRetransmits property on RTCDataChannel 
PASS RTCDataChannel interface: attribute protocol 
PASS Unscopable handled correctly for protocol property on RTCDataChannel 
PASS RTCDataChannel interface: attribute negotiated 
PASS Unscopable handled correctly for negotiated property on RTCDataChannel 
PASS RTCDataChannel interface: attribute id 
PASS Unscopable handled correctly for id property on RTCDataChannel 
FAIL RTCDataChannel interface: attribute priority assert_true: The prototype object must have a property "priority" expected true got false
PASS Unscopable handled correctly for priority property on RTCDataChannel 
PASS RTCDataChannel interface: attribute readyState 
PASS Unscopable handled correctly for readyState property on RTCDataChannel 
PASS RTCDataChannel interface: attribute bufferedAmount 
PASS Unscopable handled correctly for bufferedAmount property on RTCDataChannel 
PASS RTCDataChannel interface: attribute bufferedAmountLowThreshold 
PASS Unscopable handled correctly for bufferedAmountLowThreshold property on RTCDataChannel 
PASS RTCDataChannel interface: attribute onopen 
PASS Unscopable handled correctly for onopen property on RTCDataChannel 
PASS RTCDataChannel interface: attribute onbufferedamountlow 
PASS Unscopable handled correctly for onbufferedamountlow property on RTCDataChannel 
PASS RTCDataChannel interface: attribute onerror 
PASS Unscopable handled correctly for onerror property on RTCDataChannel 
PASS RTCDataChannel interface: attribute onclose 
PASS Unscopable handled correctly for onclose property on RTCDataChannel 
PASS RTCDataChannel interface: operation close() 
PASS Unscopable handled correctly for close() on RTCDataChannel 
PASS RTCDataChannel interface: attribute onmessage 
PASS Unscopable handled correctly for onmessage property on RTCDataChannel 
PASS RTCDataChannel interface: attribute binaryType 
PASS Unscopable handled correctly for binaryType property on RTCDataChannel 
PASS RTCDataChannel interface: operation send(USVString) 
PASS Unscopable handled correctly for send(USVString) on RTCDataChannel 
PASS RTCDataChannel interface: operation send(Blob) 
PASS Unscopable handled correctly for send(Blob) on RTCDataChannel 
PASS RTCDataChannel interface: operation send(ArrayBuffer) 
PASS Unscopable handled correctly for send(ArrayBuffer) on RTCDataChannel 
PASS RTCDataChannel interface: operation send(ArrayBufferView) 
PASS Unscopable handled correctly for send(ArrayBufferView) on RTCDataChannel 
PASS RTCDataChannel must be primary interface of new RTCPeerConnection().createDataChannel('') 
FAIL Stringification of new RTCPeerConnection().createDataChannel('') undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "label" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "ordered" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "maxPacketLifeTime" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "maxRetransmits" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "protocol" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "negotiated" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "id" with the proper type 
FAIL RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "priority" with the proper type assert_inherits: property "priority" not found in prototype chain
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "readyState" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "bufferedAmount" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "bufferedAmountLowThreshold" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onopen" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onbufferedamountlow" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onerror" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onclose" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "close()" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onmessage" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "binaryType" with the proper type 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(USVString)" with the proper type 
PASS RTCDataChannel interface: calling send(USVString) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(Blob)" with the proper type 
PASS RTCDataChannel interface: calling send(Blob) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(ArrayBuffer)" with the proper type 
PASS RTCDataChannel interface: calling send(ArrayBuffer) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError 
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(ArrayBufferView)" with the proper type 
PASS RTCDataChannel interface: calling send(ArrayBufferView) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError 
FAIL RTCDataChannelEvent interface: existence and properties of interface object undefined is not an object (evaluating 'this.array
                     .members[this.base]
                     .has_extended_attribute')
PASS RTCDataChannelEvent interface object length 
PASS RTCDataChannelEvent interface object name 
FAIL RTCDataChannelEvent interface: existence and properties of interface prototype object undefined is not an object (evaluating 'this.array
                         .members[inherit_interface]
                         .has_extended_attribute')
PASS RTCDataChannelEvent interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCDataChannelEvent interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCDataChannelEvent interface: attribute channel 
PASS Unscopable handled correctly for channel property on RTCDataChannelEvent 
PASS RTCDataChannelEvent must be primary interface of new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') }) 
FAIL Stringification of new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') }) undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')
PASS RTCDataChannelEvent interface: new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') }) must inherit property "channel" with the proper type 
FAIL RTCDTMFSender interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object length assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object name assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: operation insertDTMF(DOMString, unsigned long, unsigned long) assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
PASS Unscopable handled correctly for insertDTMF(DOMString, unsigned long, unsigned long) on RTCDTMFSender 
FAIL RTCDTMFSender interface: attribute ontonechange assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
PASS Unscopable handled correctly for ontonechange property on RTCDTMFSender 
FAIL RTCDTMFSender interface: attribute toneBuffer assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
PASS Unscopable handled correctly for toneBuffer property on RTCDTMFSender 
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object length assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object name assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: attribute tone assert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
PASS Unscopable handled correctly for tone property on RTCDTMFToneChangeEvent 
PASS RTCStatsReport interface: existence and properties of interface object 
PASS RTCStatsReport interface object length 
PASS RTCStatsReport interface object name 
PASS RTCStatsReport interface: existence and properties of interface prototype object 
PASS RTCStatsReport interface: existence and properties of interface prototype object's "constructor" property 
PASS RTCStatsReport interface: existence and properties of interface prototype object's @@unscopables property 
PASS RTCIdentityProviderGlobalScope interface: existence and properties of interface object 
PASS RTCIdentityProviderRegistrar interface: existence and properties of interface object 
FAIL RTCIdentityAssertion interface: existence and properties of interface object assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object length assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object name assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: attribute idp assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
PASS Unscopable handled correctly for idp property on RTCIdentityAssertion 
FAIL RTCIdentityAssertion interface: attribute name assert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
PASS Unscopable handled correctly for name property on RTCIdentityAssertion 
FAIL RTCErrorEvent interface: existence and properties of interface object assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object length assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object name assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: attribute error assert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
PASS Unscopable handled correctly for error property on RTCErrorEvent 
FAIL RTCErrorEvent must be primary interface of new RTCErrorEvent('error') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCErrorEvent"
FAIL Stringification of new RTCErrorEvent('error') assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCErrorEvent"
FAIL RTCErrorEvent interface: new RTCErrorEvent('error') must inherit property "error" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: RTCErrorEvent"
